open-ptc-agent/src/agent/middleware/background/counter.py
"""Tool call counter middleware for tracking subagent tool usage.

This middleware is injected into subagents to count their tool calls and
report metrics back to the BackgroundTaskRegistry.
"""

from typing import Any, Awaitable, Callable

import structlog
from langchain.agents.middleware import AgentMiddleware
from langchain_core.messages import ToolMessage
from langgraph.prebuilt.tool_node import ToolCallRequest
from langgraph.types import Command

from src.agent.middleware.background.registry import BackgroundTaskRegistry
from src.agent.middleware.background.middleware import current_background_task_id

logger = structlog.get_logger(__name__)


class ToolCallCounterMiddleware(AgentMiddleware):
    """Middleware to count tool calls and report to BackgroundTaskRegistry.

    This middleware is designed to be injected into subagents running in the
    background. It tracks how many tool calls each subagent makes and what
    tools are being used.

    The middleware uses a contextvar (current_background_task_id) to identify
    which background task it belongs to. Contextvars properly propagate across
    await boundaries, ensuring tool calls are tracked even when subagents
    execute in different execution contexts.

    Usage:
        # Create counter middleware with shared registry
        counter = ToolCallCounterMiddleware(registry=background_middleware.registry)

        # Inject into subagent specs
        subagent_spec["middleware"] = [counter]
    """

    def __init__(self, registry: BackgroundTaskRegistry):
        """Initialize the counter middleware.

        Args:
            registry: The BackgroundTaskRegistry to report metrics to
        """
        super().__init__()
        self.tools = []  # No additional tools
        self.registry = registry

    def wrap_tool_call(
        self,
        request: ToolCallRequest,
        handler: Callable[[ToolCallRequest], ToolMessage | Command],
    ) -> ToolMessage | Command:
        """Synchronous wrap_tool_call - no tracking in sync mode."""
        return handler(request)

    async def awrap_tool_call(
        self,
        request: ToolCallRequest,
        handler: Callable[[ToolCallRequest], Awaitable[ToolMessage | Command]],
    ) -> ToolMessage | Command:
        """Count tool call and report to registry.

        This method:
        1. Extracts the tool name from the request
        2. Gets the task_id from the asyncio task context
        3. Reports the metric to the registry
        4. Executes the tool call
        """
        # Extract tool name
        tool_call = request.tool_call
        tool_name = tool_call.get("name", "unknown")

        # Get task_id from contextvar (set by BackgroundSubagentMiddleware)
        # Contextvars properly propagate across await boundaries
        task_id = current_background_task_id.get()

        # Report metric to registry before execution
        if task_id and self.registry:
            await self.registry.update_metrics(task_id, tool_name)
            logger.debug(
                "Counted tool call for background task",
                task_id=task_id,
                tool_name=tool_name,
            )

        # Execute the tool call
        return await handler(request)
